fix: User SEQ 타입 통일 및 JWT subject 조회 로직 수정
- JWT subject에 저장된 SEQ로 회원 조회하도록 수정
- User SEQ 타입을 Long에서 Integer로 통일 (DB INT 타입과 일치)
- getUserIdBySeq() 메서드 추가 (SEQ → userId 변환)

+++ src/main/java/kr/wisead/domain/user/service/UserService.java
@@ -468,7 +468,7 @@ public class UserService {
    * @param seqList 삭제할 회원 시퀀스 목록
    */
   @Transactional
-  public void deleteUsers(List<Long> seqList) {
+  public void deleteUsers(List<Integer> seqList) {
     if (seqList == null || seqList.isEmpty()) {
       throw new BusinessException(ErrorCode.INVALID_INPUT_VALUE, "삭제할 회원 목록이 비어있습니다.");
     }
